vue/vue3-playground/python/requirements.txt
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.2
click==8.3.1
cramjam==2.11.0
et_xmlfile==2.0.0
geobuf==2.0.0
h11==0.16.0
idna==3.10
numpy==2.3.0
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
protobuf==6.33.2
PyPDF2 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_e9jdhxarzu/croot/pypdf2_1733841481306/work
PySocks==1.7.1
python-dotenv==1.1.0
python-snappy==0.7.3
pyupyun==0.1.3
requests==2.32.4
scipy==1.15.3
selenium==4.38.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wheel==0.45.1
wsproto==1.2.0
